used reselect for filter, changed Grid and List back to components, erased App local state
